WintunCreateAdapter issures
